Separate file for GridChildShape.
